ai-backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
torch==2.1.0
torchvision==0.16.0
diffusers==0.25.0
transformers==4.36.0
accelerate==0.25.0
huggingface-hub==0.20.3
controlnet-aux==0.0.7
pillow==10.1.0
opencv-python==4.8.1.78
scipy==1.11.4
safetensors==0.4.1
omegaconf==2.3.0
einops==0.7.0
